refactor(ui): split into modules
